api/requirements.txt
fastapi
uvicorn[standard]
duckdb
pydantic
python-dotenv
